+++ activerecord/test/cases/calculations_test.rb
@@ -130,6 +130,7 @@ class CalculationsTest < ActiveRecord::TestCase
       9 => 53
     }
     assert_equal expected, accounts.sum(:credit_limit)
+    assert_equal expected, accounts.merge!(accounts).uniq!(:group).sum(:credit_limit)
 
     expected = {
       [nil, nil] => 50,
@@ -138,7 +139,14 @@ class CalculationsTest < ActiveRecord::TestCase
       [6, 6] => 55,
       [9, 9] => 53
     }
-    assert_equal expected, accounts.merge!(accounts).maximum(:credit_limit)
+    message = <<-MSG.squish
+      `maximum` with group by duplicated fields does no longer affect to result in Rails 6.2.
+      To migrate to Rails 6.2's behavior, use `uniq!(:group)` to deduplicate group fields
+      (`accounts.uniq!(:group).maximum(:credit_limit)`).
+    MSG
+    assert_deprecated(message) do
+      assert_equal expected, accounts.merge!(accounts).maximum(:credit_limit)
+    end
 
     expected = {
       [nil, nil, nil, nil] => 50,
@@ -147,7 +155,14 @@ class CalculationsTest < ActiveRecord::TestCase
       [6, 6, 6, 6] => 50,
       [9, 9, 9, 9] => 53
     }
-    assert_equal expected, accounts.merge!(accounts).minimum(:credit_limit)
+    message = <<-MSG.squish
+      `minimum` with group by duplicated fields does no longer affect to result in Rails 6.2.
+      To migrate to Rails 6.2's behavior, use `uniq!(:group)` to deduplicate group fields
+      (`accounts.uniq!(:group).minimum(:credit_limit)`).
+    MSG
+    assert_deprecated(message) do
+      assert_equal expected, accounts.merge!(accounts).minimum(:credit_limit)
+    end
   end
 
   def test_should_generate_valid_sql_with_joins_and_group